requirements.txt
click==6.7
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Flask-Bcrypt==0.7.1
bcrypt==3.1.4
Werkzeug==0.14.1
WTForms==2.2.1
# Flask-Login
Flask-Script
Flask-Migrate
Flask-SQLAlchmey
gunicorn
Flask-JWT
PyJWT==1.6.1

